// src/lip/types/product.ts
import { ProductCollection, 
    ProductSize, 
    ProductStatus }
    from "../enums/product.enum";

export interface Product {
    _id: string,
    productStatus: ProductStatus;
    productCollection:  ProductCollection;
    productName: string;
    productPrice: number;
    productLeftCount: number;
    productSize: ProductSize;
    productVolume: number;
    productDesc?: string;
    productImage: string[];
    productViews: number;
    createdAt: Date;
    updatedAt: Date;
}

export interface ProductInquery {
    order: string,
    page: number,
    limit: number,
    productCollection?: ProductCollection,
    search?: string  
}